Reorganize code and tests
