+++ src/test/scala/io/qbeast/spark/utils/QbeastFilterPushdownTest.scala
@@ -1,11 +1,11 @@
 package io.qbeast.spark.utils
 
+import io.qbeast.TestUtils.{checkFileFiltering, checkLogicalFilterPushdown}
 import io.qbeast.spark.QbeastIntegrationTestSpec
 import io.qbeast.spark.delta.OTreeIndex
-import io.qbeast.spark.internal.expressions.QbeastMurmur3Hash
-import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.execution.FileSourceScanExec
 import org.apache.spark.sql.functions.{avg, col, rand, regexp_replace, when}
+import org.scalatest.exceptions.TestFailedException
 
 class QbeastFilterPushdownTest extends QbeastIntegrationTestSpec {
 
@@ -15,40 +15,6 @@ class QbeastFilterPushdownTest extends QbeastIntegrationTestSpec {
   private val filter_product_greaterThanOrEq = "(product_id < 50500010)"
   private val filter_user_equal = "(user_id = 536764969)"
   private val filter_product_equal = "(product_id = 11522682)"
-
-  private def checkFiltersArePushedDown(query: DataFrame): Unit = {
-    val leaves =
-      query.queryExecution.executedPlan.collectLeaves().filter(_.isInstanceOf[FileSourceScanExec])
-
-    leaves should not be empty
-
-    leaves.exists(p =>
-      p
-        .asInstanceOf[FileSourceScanExec]
-        .relation
-        .location
-        .isInstanceOf[OTreeIndex]) shouldBe true
-
-    leaves
-      .foreach {
-        case f: FileSourceScanExec if f.relation.location.isInstanceOf[OTreeIndex] =>
-          f.dataFilters.nonEmpty shouldBe true
-      }
-  }
-
-  private def checkLogicalFilterPushdown(sqlFilters: Seq[String], query: DataFrame): Unit = {
-    val leaves = query.queryExecution.sparkPlan.collectLeaves()
-
-    val dataFilters = leaves
-      .collectFirst {
-        case f: FileSourceScanExec if f.relation.location.isInstanceOf[OTreeIndex] =>
-          f.dataFilters.filterNot(_.isInstanceOf[QbeastMurmur3Hash])
-      }
-      .getOrElse(Seq.empty)
-
-    val dataFiltersSql = dataFilters.map(_.sql)
-    sqlFilters.foreach(filter => dataFiltersSql should contain(filter))
-  }
 
   "Qbeast" should
     "return a valid filtering of the original dataset " +
